Fix Exercise 6 test
The acceptance test of Exercise 6 used the `href` attribute to find the element we want to click on, this test gets broken when in the next exercises we change the URL to use the `slug` property instead of the course `title`. In this change I recommend to use the `data-name` attribute, which remains unchanged throughout the duration of the training.